+++ src/news/repository.py
@@ -30,24 +30,27 @@ class NewsRepository(AbstractRepository):
         return len(news)
 
     def delete(self, news_id: int):
-        news = self.db.session.query(News).filter(News.id == news_id).first()
+        news = self.db.session.query(News).filter(News.id == news_id).one_or_none()
+        if news is None:
+            raise KeyError("news not found")
         self.db.session.delete(news)
         self.db.session.commit()
 
     def update(self, news_data: NewsDto):
-        news = self.db.session.query(News).filter(News.id == news_data.id).first()
-        news.title = news_data.title
+        news = self.db.session.query(News).filter(News.id == news_data.id).one_or_none()
+        if news is None:
+            raise KeyError("news not found")
         news.content = news_data.content
         news.image_url = news_data.image_url
 
         self.db.session.add(news)
         self.db.session.commit()
 
-    def get(self, news_id: int | None = None, title: str | None = None) -> NewsDto:
+    def get(
+        self, news_id: int | None = None
+    ) -> NewsDto | None:
         if news_id:
-            return self.db.session.query(News).filter(News.id == news_id).first()
-        elif title:
-            return self.db.session.query(News).filter(News.title == title).first()
+            return self.db.session.query(News).filter(News.id == news_id).one_or_none()
         else:
             raise ValueError("news_id or title must be specified")
